requirements.api.txt
fastapi
uvicorn
groq
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv
requests